경매 등록 기능 완성 & attachment, photo, category  관련 백엔드 기반 작업 완성

+++ auctionara/src/main/java/com/an/auctionara/entity/AttachmentDto.java
@@ -11,5 +11,5 @@ public class AttachmentDto {
 	private String attachmentUploadname;
 	private String attachmentSavename;
 	private String attachmentType;
-	private int attachmentSize; 
+	private long attachmentSize; 
 }
